pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.0"
description = "Trying out some thing is grid2op."
authors = [ "Viktor Todosijevic <todosijevicviktor998@gmail.com>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
hydra-core = "^1.3.2"
grid2op = {extras = ["optional"], version = "^1.10.5"}
pandapower = "^2.14.11"
numba = "^0.61.0"
matplotlib = "3.7.0"
ipympl = "^0.9.7"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pylint = "^3.0.3"
black = "^23.12.1"
isort = "^5.13.2"
bump-my-version = "^0.16.2"
flynt = "^1.0.1"
flake8 = "^7.0.0"
pre-commit = "^3.7.0"
flake8-pyproject = "^1.2.3"
docformatter = {extras = ["tomli"], version = "^1.7.5"}
nox = "^2024.4.15"
mccabe = "^0.7.0"

[tool.poetry.group.lab]
optional = true

[tool.poetry.group.lab.dependencies]
jupyterlab = "^4.0.10"
notebook = "^7.0.6"
jupyterlab-lsp = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#
# Scripts
#
[tool.poetry.scripts]

#
# Tools
#

[tool.bumpversion]
current_version = "0.0.0"
commit = true
message = "Version {new_version} released."
tag = true
tag_name = "{new_version}"
parse = '(?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+) (?P<releaseTime>.*)?'
serialize = [
    "{major}.{minor}.{patch}",
    "{utcnow:%Y-%m-%dT%H:%M:%SZ}"
]

[[tool.bumpversion.files]]
filename = "CHANGES.md"
search = '''
[Unreleased](<REPOSITORY_URL>/tree/main) (latest)
-------------------------------------------------------------------------------------
'''
replace = '''
[Unreleased](<REPOSITORY_URL>/tree/main) (latest)
-------------------------------------------------------------------------------------

[//]: # (New changes here in list form)

[{new_version}](<REPOSITORY_URL>/tree/{new_version}) ({utcnow:%Y-%m-%d})
-------------------------------------------------------------------------------------
'''

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "Makefile.variables"
search = "APP_VERSION := {current_version}"
replace = "APP_VERSION := {new_version}"

[tool.pylint]
disable = "C0114,C0115,C0116,R0903,R1710,W1203"
max-line-length = 120
max-locals = 20
max-args = 16
max-attributes = 20
ignore = [
    ".git",
    "migrations",
    "__pycache__"
]

[tool.flake8]
max-line-length = 120
ignore = ["E203", "E266", "E501", "W503"]
docstring-convention = "numpy"
max-complexity = 20
per-file-ignores = []
exclude = [
    ".git",
    "migrations",
    "__pycache__"
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.flynt]
line-length = 120
transform-concats = true
verbose = true

[tool.docformatter]
style = "numpy"
pre-summary-newline = true
wrap-descriptions = 120
wrap-summaries = 120
blank = false
exclude = []
recursive = true